Improve setup script

+++ setup.sh
@@ -1,1 +1,3 @@
-cp .env.development.schema .env.development
+if [ ! -f ./.env ]; then
+  cp .env.development.schema .env.development
+fi
